Rearrange partners cron workflows
