+++ apps/web/lib/partners/complete-program-applications.ts
@@ -197,7 +197,7 @@ export async function completeProgramApplications(userEmail: string) {
         // Auto-approve the partner if the group has auto-approval enabled
         group?.autoApprovePartnersEnabledAt
           ? qstash.publishJSON({
-              url: `${APP_DOMAIN_WITH_NGROK}/api/cron/auto-approve-partner`,
+              url: `${APP_DOMAIN_WITH_NGROK}/api/cron/partners/auto-approve`,
               delay: 5 * 60,
               body: {
                 programId: program.id,
